% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.R
\name{mean}
\alias{mean}
\title{mean}
\usage{
mean(x)
}
\arguments{
\item{x}{a vector of numbers that we will use to take mean}

\item{s}{The lower trim bound}

\item{l}{The upper trim bound}
}
\value{
Returns a trimmed mean
}
\description{
This function will ...
}
\examples{
mean(1:5)


}
\author{
Leean Jo
}
