% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p.R
\name{p}
\alias{p}
\title{p
The second section is a longer description of the function.
This can go on for multiple lines.}
\usage{
p(a, b, c)
}
\arguments{
\item{x}{numeric vector}
}
\description{
p
The second section is a longer description of the function.
This can go on for multiple lines.
}
