% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pythagorean.R
\name{pythagorean}
\alias{pythagorean}
\title{pythagorean
The second section is a longer description of the function.
This can go on for multiple lines.}
\usage{
pythagorean(x, y)
}
\arguments{
\item{x}{numeric vector}
}
\description{
pythagorean
The second section is a longer description of the function.
This can go on for multiple lines.
}
