% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_mean.R
\docType{data}
\name{x}
\alias{x}
\title{t_mean}
\format{
An object of class \code{numeric} of length 8.
}
\usage{
x
}
\arguments{
\item{x}{a vector of numbers that we will use to take mean}

\item{s}{The lower trim bound}

\item{fl}{The first upper trim bound}

\item{sl}{The second upper trim bound}
}
\value{
Returns a trimmed mean
}
\description{
This function will ...
}
\examples{
t_mean(1:5)


}
\author{
Leean Jo
}
\keyword{datasets}
